// interfaces/historial.ts
export interface HistorialInterface {
  _id: string;
  priorOrg: string;
  etapaOrg: string;
  estadoOrg: string;
  vendedorOrg: string;
  diseniadorOrg: string;
  priorAct: string;
  etapaAct: string;
  estadoAct: string;
  vendedorAct: string;
  diseniadorAct: string;
  idPedido: string;
  fecha: string;
  usuario: string;
  foranea: string;
}